Add permission checks to avoid crashes

+++ app/src/main/java/com/bnyro/contacts/ui/models/ContactsModel.kt
@@ -1,44 +1,37 @@
 package com.bnyro.contacts.ui.models
 
 import android.Manifest
+import android.annotation.SuppressLint
 import android.content.Context
-import android.content.pm.PackageManager
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
-import androidx.core.app.ActivityCompat
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.bnyro.contacts.MainActivity
 import com.bnyro.contacts.obj.ContactData
 import com.bnyro.contacts.util.ContactsHelper
+import com.bnyro.contacts.util.PermissionHelper
 import kotlinx.coroutines.launch
 
 class ContactsModel : ViewModel() {
     var contacts by mutableStateOf<List<ContactData>?>(null)
-    var contactsHelper: ContactsHelper? = null
+    private var contactsHelper: ContactsHelper? = null
 
+    @SuppressLint("MissingPermission")
     fun loadContacts(context: Context) {
         viewModelScope.launch {
             contactsHelper = ContactsHelper(context)
-            if (ActivityCompat.checkSelfPermission(
-                    context,
-                    Manifest.permission.READ_CONTACTS
-                ) != PackageManager.PERMISSION_GRANTED
-            ) {
-                ActivityCompat.requestPermissions(
-                    context as MainActivity,
-                    arrayOf(Manifest.permission.READ_CONTACTS, Manifest.permission.WRITE_CONTACTS),
-                    1
-                )
-                return@launch
-            }
+            if (!PermissionHelper.checkPermissions(context, Manifest.permission.READ_CONTACTS)) return@launch
             contacts = contactsHelper!!.getContactList()
         }
     }
 
-    fun deleteContact(contact: ContactData) {
-        contactsHelper?.deleteContacts(listOf(contact))
-        contacts = contacts?.minus(contact)
+    @SuppressLint("MissingPermission")
+    fun deleteContact(context: Context, contact: ContactData) {
+        viewModelScope.launch {
+            if (!PermissionHelper.checkPermissions(context, Manifest.permission.READ_CONTACTS)) return@launch
+            contactsHelper?.deleteContacts(listOf(contact))
+            contacts = contacts?.minus(contact)
+        }
     }
 }
